pyproject.toml
[tool.poetry]
name = "vanutp-music-bot"
version = "0.1.0"
description = ""
authors = ["Ivan Filipenkov <hello@vanutp.dev>"]

[tool.poetry.dependencies]
python = "^3.11"
telethon-v1-24 = "^1.24.14"
cryptg = "^0.4.0"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
yt-dlp = "^2023.12.30"
aiorun = "^2023.7.2"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
